Replace per-step `RUSTFLAGS` with global one
This should not make any difference for this workflow, since all steps
already set the flag. But it should make the workflow easier to maintain
going forward. I think failing the CI build on warnings is always a good
idea, and now we don't have to remember that for each new step.